requirements.txt
flask>=2.3.0
requests>=2.31.0
pdf2image>=1.17.0
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-docx>=1.1.0
docx2pdf>=0.1.8
pywin32>=308

